requirements.txt
streamlit
llama-index
firecrawl-py
httpx
python-dotenv
openai 